feat: change to archi 1.0.0 for SmartRecruiters
