+++ src/hrflow_connectors/connectors/__init__.py
@@ -1,1 +1,2 @@
 from .crosstalent import *
+from .smartrecruiters import *
