--- src/hrflow_connectors/connectors/boards/smartrecruiters/__init__.py
@@ -1,2 +1,0 @@
-from .actions import *
-from .spec import *
