--- src/hrflow_connectors/connectors/boards/smartrecruiters/spec.py
@@ -1,8 +1,0 @@
-from pydantic import BaseModel
-from typing import Tuple
-
-from .actions import SmartRecruitersPullJobsAction
-
-
-class Spec(BaseModel):
-    actions: Tuple[SmartRecruitersPullJobsAction]
