--- src/hrflow_connectors/connectors/destinations/smartrecruiters/__init__.py
@@ -1,2 +1,0 @@
-from .actions import *
-from .spec import *
