--- src/hrflow_connectors/connectors/destinations/smartrecruiters/spec.py
@@ -1,8 +1,0 @@
-from pydantic import BaseModel
-from typing import Tuple
-
-from .actions import SmartRecruitersPushProfileAction
-
-
-class Spec(BaseModel):
-    actions: Tuple[SmartRecruitersPushProfileAction]
