+++ src/hrflow_connectors/connectors/smartrecruiters/README.md
@@ -1,0 +1,6 @@
+# Smart Recruiters Connector
+**Smart Recruiters is an ATS.**
+
+Lists of actions :
+* [**Pull jobs**](docs/pull_jobs.md)
+* [**Push profile**](docs/push_profile.md)
