+++ src/hrflow_connectors/connectors/smartrecruiters/__init__.py
@@ -1,0 +1,1 @@
+from .connector import *
