+++ tests/connectors/smartrecruiters/test_pull_jobs.py
@@ -1,7 +1,7 @@
 import pytest
 
-from hrflow_connectors.core.auth import XSmartTokenAuth
-from hrflow_connectors.connectors.boards.smartrecruiters import SmartRecruitersPullJobsAction
+from hrflow_connectors import XSmartTokenAuth
+from hrflow_connectors import SmartRecruiters
 
 
 @pytest.fixture
@@ -12,8 +12,8 @@ def auth(credentials):
     return auth
 
 
-def test_GetAllJobs(logger, auth, hrflow_client):
-    action = SmartRecruitersPullJobsAction(
+def test_PullJobsAction(logger, auth, hrflow_client):
+    SmartRecruiters.pull_jobs(
         auth=auth,
         hrflow_client=hrflow_client("dev-demo"),
         limit=2,
@@ -22,4 +22,3 @@ def test_GetAllJobs(logger, auth, hrflow_client):
         archive_deleted_jobs_from_stream=False,
         posting_status="PUBLIC",
     )
-    action.execute()
